fix of wrong status change processing, caused by a patch for #1304
fixes #1321 (When connecting to only some protocols, the notification area icon shows offline icon)

+++ plugins/Clist_nicer/src/cluiservices.cpp
@@ -173,7 +173,7 @@ void CluiProtocolStatusChanged(int, const char*)
 		char *szMaxProto = nullptr;
 		wStatus = Clist_GetGeneralizedStatus(&szMaxProto);
 		iIcon = IconFromStatusMode(szMaxProto, wStatus, 0);
-		g_maxStatus = (int)wStatus;
+		g_maxStatus = wStatus;
 	}
 
 	/*
